requirements.txt
beautifulsoup4
Flask
Jinja2
lxml
MarkupSafe
requests
urllib3
Werkzeug
bs4
gunicorn
flask-cors
Pillow
pytesseract
selenium